changed edit to a modal
